src/requirements.txt
absl-py==0.9.0
actionlib==1.12.0
angles==1.9.11
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blessings==1.7
bondpy==1.8.3
cachetools==4.1.1
camera-calibration==1.13.0
camera-calibration-parsers==1.11.13
carla-ackermann-control==0.0.0
carla-ad-agent==0.0.1
carla-ego-vehicle==0.0.0
carla-infrastructure==0.0.0
carla-manual-control==0.0.0
carla-ros-bridge==0.0.1
carla-ros-scenario-runner==0.0.0
carla-spectator-camera==0.0.0
carla-twist-to-control==0.0.0
carla-waypoint-publisher==0.0.0
catkin==0.7.17
certifi==2020.6.20
chardet==3.0.4
controller-manager==0.15.1
controller-manager-msgs==0.15.1
cpools==0.0.0
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
dynamic-reconfigure==1.6.0
entrypoints==0.3
et-xmlfile==1.0.1
gast==0.3.3
gazebo-plugins==2.8.4
gazebo-ros==2.8.4
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.8
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
image-geometry==1.13.0
importlib-metadata==1.7.0
interactive-markers==1.11.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kdl-parser-py==1.13.1
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
keras-tcn==3.1.1
kiwisolver==1.2.0
laser-geometry==1.6.4
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
message-filters==1.14.3
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
open3d-python==0.7.0.0
opencv-python==4.4.0.42
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySimpleGUI==4.28.0
python-dateutil==2.8.1
python-igraph==0.8.2
python-qt-binding==0.3.5
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.24.0
requests-oauthlib==1.3.0
resource-retriever==1.12.5
rosbag==1.14.3
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.6
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.6
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.6
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-carla-control==0.0.0
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.10
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rsa==4.6
rviz==1.13.4
scipy @ file:///tmp/build/80754af9/scipy_1592930510235/work
Send2Trash==1.5.0
sensor-msgs==1.12.7
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.3
tornado==6.0.4
traitlets==4.3.3
urdfdom-py==0.4.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xacro==1.13.3
zipp==3.1.0
